Remove object references from isAvailableForFileTest case names
The default toString used in the parameterized test case names includes
object references for the FileType. Since these vary every run, test
analyzers consider them a different test every time, which is
undesirable.

However, this is easy to fix with the existing setup. We just need to
override toString for FileType to print fixed string per type. FileType
already has a human readable `description` we can use for this purpose.
This also requires us to mock out `description` for a mock FileType,
which would be `null` otherwise. We do this by pulling it into a private
helper factory function for readability.

Fixes: 232562525
Test: N/A
Change-Id: Ib2ccf70c59400c1925dcbb45a461e30fdfb160d4