Create PanelAction.kt as an abstract base class for actions that use the StringResourceViewPanel.
Currently many AnAction classes hold a reference to a StringResourceViewPanel. However,
AnAction classes should be stateless according to
https://plugins.jetbrains.com/docs/intellij/basic-action-system.html#action-implementation.
The Panel however should be accessible via the AnActionEvent that is passed to both the
update and actionPerformed methods. This base class makes accessing the panel that way
significantly easier and will make refactoring those classes for testability significantly
easier - we can just inject the Panel via the AnActionEvent instead of needing to construct
one to instantiate the AnAction.

Bug: 232444069
Test: Tested by PanelActionTest.kt

Change-Id: I20348d260b2d2583999945372ecc29c5ddf58f79