Move StringsWriteUtils.createItem into DalUtils.java which is the only place it is used.
This method is not used in the Translations Editor so clearly has no
business being in here. Instead, move it to the one place it is used
where it can become an implementation detail of an already-tested
method.

Remove the tests for this method as it is no longer part of
the public API of StringsWriteUtils.

Bug: 232444069
Test: Existing unit/presubmit tests.
Change-Id: Ic7658c21da92371291b0e57eaa2ff23a89338dce