Include task finder results in project dumps
This is to track any changes to tasks we run, to make it possible to
simplify the `GradleTaskFinder`, to highlight issues with support for
composite builds.

We dump task finder results in all `BuildMode` and for all
`TestCompileMode`s. However, since currently there are several ways of
getting tasks to run in test modes we include both result from
`taskFinder.findTasksToExecute` and results from
`GradleModuleTasksProvider.getUnitTestTasks` (for the unit test mode) or
`GradleModuleTasksProvider.getTasksFor` (which simply calls
`findTasksToExecute`) as these are the ways tasks are retrieved in AS.

The `GradleModuleTasksProvider` output is only reported as a diff to
the regular task finder output.

Also: Ensure stable order of tasks returned by
      `GradleModuleTasksProvider.getUnitTestTasks`.

Bug: n/a
Test: n/a
Change-Id: If959b4e4276b7b7958ec85fd6c1030c0237d73d0