Fix VectorDrawableTransformer
The existing code seemed to have a bug when computing the values
for translation when the size aspect ratio was different from the
viewport aspect ratio.

The code is now changed such that the viewport is not changed,
and the scale is applied such that the shape of the image is not
distorted if the target aspect ration is different from the view
aspect ratio.

I added a clip-path when this method is used in the Resource
Manager because a larger display area could cause a part of an
image to show that would otherwise be clipped. This seems to
happen when a transformation is applied to the original image.

Bug: 173833413
Test: Unit test added
Change-Id: Id56ad4114b1cda9d9938a51e682dd8ae5bc8c874