Use a Single Logger All Over the Logcat Module
For some reason, the common practice is to use thisLogger() which uses the classname as a category.

There are several problems with this pattern. The most important one is that it's hard to find the logs for a specific module. The package name usually doesn't get logged completely, just the class name so you have to filter with all the classes you are interested in.

Using a common logger for a module makes it easy to find the relevant logs.

Bug: n/a
Test: n/a
Change-Id: Ide406894c8ffd9609d7c6eb7ec70cae97a852b68