Fix module hierarchy in Android project view
Module names given by the IntelliJ platform to Gradle modules usually
represent the hierarchy of projects in Gradle build configuration but
this is not always the case. When it happens that the platform cannot
create a module with a desired name it applies various disambiguation
tactics and breaks the hierarchy. The same may happen to any modules
representing Gradle projects whose root directory is configured to be
outside of the IDE project root.

This change makes sure that the hierarchy in the Android project view
is always based on the Gradle configuration.

Also: fix kotlinMultiplatform test project (add missing `kotlin` folder)

Bug: n/a
Test: AndroidProjectViewSnapshotComparisonTestDef
Change-Id: If8ab9c4bc1ce6bf57108d80e6196ff79b93cf3f1