Use the screen size instead of View size for zoom to fit
In some apps there may be multiple root views e.g. when a dialog or
popup is open there will be multiple DecorView instances.

Normally a dialog or popup is opened as a result of a user action
in the app. But if an app should start with an open dialog/popup
then we may zoom to fit the dialog or popup instead of the main
screen.

It is confusing when we zoom to fit the view size that is different
from the screen size.

Use the new screen size passed from the agent to handle zoom to fit.

Fixes: 232987010
Test: Add check of zoom from screen size
Change-Id: Ib1843409de0aed7b02e24ee4098fbe7ad88ec77c