[AUA] Rearrange ToolWindow View initialization
This allows a more obviously correct (and also a more correct)
initialization of the UI elements from the model, which previously
could not be done because of needing to refer to elements of the view
before their instantiation was complete.

This unblocks adding some tests for initial UI states (before any
direct user interaction).

Bug: n/a
Test: added
Change-Id: Ia99d1eb79b3f63cc8f9f07715ce2e1081f602352