Fixes race condition in Xvfb
There was a race condition where the Xvfb process was
running but was not ready to accept connections.

Newer versions of Xvfb have the option -displayfd,
to print in a fd when they are ready. However we don't
have that.

There is a way of knowing when the server is ready
and it's monitoring the domain socket that it opens.
That socket is located in /tmp/.X11-unix/X1234

This change does that and fully eliminates the race
condition. Validated with StartUp test, that went
from 50/500 failures to 0/500.

Test: this is it
Fixes: 232081272
Change-Id: I73d401ab434a796c5bf7a2bfc8ab5285b0fa9f0e