[AUA] Improve test handling of unshown dialogs
If they're not shown or interacted with (even headlessly), then the
headless operation to dispose them on close can never happen.  Many of
our tests do not close their dialogs, instead disposing them
explicitly, but it is easy to fail to do so.  Move to registering them
to be disposed on tearDown, and add a check to verify that there are
no undisposed objects associated with DialogWrappers after tearDown
has run.

Bug: none filed
Test: existing (no production change)
Change-Id: I4078cbd77a9161a8560f95436a79d1af10602abe