Show shortcut key in the tooltips of zoom action
For the registered AnAction class, even if it is an object class,
Intellij framework still use reflection to create the instance.
This makes the ZoomAction instance in DesignSurface are different than
the instances in ActionManager.

This change make design tool always use the ZoomAction instances in the
ActionManager, which have the shortcuts and will display it in their
tooltips.

Fixes: 232720346
Test: N/A (UI-tweak)
Change-Id: I8d7b57036af13de3e244123eb78a7f7969d25815