Remove redundant logic from AndroidComplicationConfigurationEditor
In AndroidComplicationConfigurationEditor we had originalChosenSlots variable and custom isModified implementation for slots component. But these things are unnecessary in real life.
To understand whether settings are modified platform compares serialized versions of original settings and settings after invocation of "applyTo". So having originalChosenSlots and corresponding logic could lead to wrong assumptions that we affect on state of "Apply" button with them.

I removed this logic and rewrote tests in a way that rely on the platform mechanism of resetting/applying/checking isModified logic.

Bug: 237507678
Test: AndroidComplicationConfigurationEditorTest is amended
Change-Id: I518ffc79485104833144e7e8736dc66bd64b55ba