Add completion popup to the ChangeManager
This delays the triggering of updates until the completion popup is
closed. This ensures that unecessary completion changes are not
triggered when we know that additional changes might be coming
(completion).

This logic used to be specific to the Compose Preview. Moving it to the
generic editor change handling for Designers.

Test: ChangeManagerTest
Bug: N/A
Change-Id: I82e78e968c3c65584d1aafc8d4d091487f560224