Allow test scopes to correctly include R files
This change adjusts isLightVirtualFileFromAccessibleModule to allow
scopes for test and android test modules to be able to access the R
class of the corrisponding main module. This was broken when we switched
to creating one module per android artifact as this method assumed
visiblity within modules was the same as within Gradle projects.

Fixes: 232001921
Bug: 235374639
Test: test added
Change-Id: I711199a1b9e23a2b129688799487a8151d047d97