Introduces AndroidSystem to the test framework
This class represents the system the user is running
studio on. It contains the file system, the display,
the environment variables, etc.

It provides an abstraction of the common operations
that are done on the system to ease the writing of tests.

Note that complicated tests, like UpdateTest, still
can use the raw apis for more complex flows.

Test: these
Bug: NA
Change-Id: I5c987ee26af2d9feb877c66e7e701de29b564424