[PSD] Set preferred size of issuesView in dependencies
... IssueViewePanel so it matches the preferred size of its
containing panel.

Currently there is no preferred size specified and this is causing
different behaviour in Linux and Mac. For Linux the panel is
resized to fit the content and thus the issues are shown, but in
Mac, the pane is not resized and thus no issues are displayed when
two or more type of issues are present.

This change sets the preferred size of the issuesView panes to
macth the preferred size of the containing panel and in this way
the panes are resized to fit the content even on Mac.

Bug: 238476607
Test: Manually confirmed the content is visible in Mac and Linux.
Change-Id: I65cb93c6e0471ad8b5a10a3029defb1fadf5ce56