Move Compose Preview warning icon
Each Compose Preview contains a status icon in its top-left corner.
However, when a render error occurs when rendering a Preview, such
status icon disappears and a warning icon is shown in the top-right
corner showing that something went wrong.

This change moves the behaviour of the warning icon mentioned above
into the status icon that is present in the top-left corner of each
preview.

Bug: 232716935
Bug: 239077332
Test: ComposePreviewStatusIconActionTest
Change-Id: Id3dfaa28bff80b20db77e798fd0f559996135856