[Dsl] Preserve file access order in GradleDslFileCache
Most Dsl operations are not sensitive to the order that entries in the
cache are accessed.  However, a request to reparse the Dsl,
implemented by clearing and re-parsing the files in the
GradleDslFileCache, is order-sensitive: if the files are re-parsed in
the wrong order, there may be dangling references and injections to
the previous parsed contents.  (This is particularly important for
cross-file references, which become substantially more common when we
support Gradle Version Catalogs).

Fixes: 224949529
Test: added
Change-Id: I19953263c64491ecb54a79d593f27f87ea5667c2