Gracefully shut down Studio
The gradle daemon is still alive after we kill (not shut down)
Studio. This allows the daemon to continue to write to the
filesystem, thus causing a race in the shut down's deletion of
directories. I.e. shutdown would clean out a folder, but gradle
would create files in the folder before the folder itself can
be deleted.

By gracefully shutting down Studio, we are able to get Studio
to shut down the gradle daemon as well.

Fixes: 239343337
Test: test that was failing now passes 500/500
Change-Id: Ic2e935f3bc458230377273eefb71af27d5e6ea77