[DeviceSpec] Prioritize user dimensions when creating Device
When creating a Device instance for the Preview, there's a risk of
introducing an error if the dpi doesn't match any of the predefined
Density values (ldpi, mdpi, hdpi, etc.).

Before, error would always be transferred to the dimensions in 'dp' so
that the Device would match the DeviceSpec dimension in pixels.

This makes it so that the Device matches the user defined dimensions in
the DeviceSpec, matching both value and unit. So the Device matches the
dimensions in 'dp' if that's how the user defined it, and vice-versa.

Fixes: 239816491
Test: updated
Change-Id: Ifc501afc1e7da921d764247c047db62514fd5aac