Avoid ProcessCancelledException breaking tests
There is a race introduced by a workaround for b/171914220. It is still
not clear why would checkCancelled not interrupt execution with a write
action pending under `ProgressIndicatorUtils.runWithWriteActionPriority`

The workaround may result in ProcessCancelledException being thrown if
a write action is submitted just after `.checkCancelled()` completed.
We can fix this problem without removing the workaround by making sure
the current progress indicator is cancelled as it makes the framework
to handle the ProcessCancelledException as cancellation.

Bug: 235869541
Bug: 224542619
Test: n/a
Change-Id: I9d50ea6fba4933fcc525a16fb3859a83dbef90c8