Timeout requests in SyncLayoutlibSceneManager
This sets a timeout of 60 seconds for operations on the
SyncLayoutlibManager to avoid Bazel timing out the test cases when
blocked.

If Bazel timesout the test, it will fail a number of other tests making
it hard to debug. This ensures that the test receiving the timeout is
flagged without Bazel killing it.

Test: N/A
Bug: N/A
Change-Id: I563247a1682d1a087d0a9ea9fa467fb093620413