Don't download files if DBI is not the selected tab in AppInspection
Currently after DBI is started, it tries to enter offline mode as soon
as the device goes offline. Even if it's not visible.

This was ok originally when it was the only inspector. But now
AppInspection has multiple tabs for different inspectors, so the
user is not necessarily using DBI when interacting with AppInspection.

Test: OfflineModeManagerTest.kt
Bug: 239816903
Change-Id: I8b262f5fc99ce85b8491c619187f228f26f12a40