Fix the slider action
The custom slider control is used for:
- layout spacing
- alpha overlay value

Both these slider do not work when the toolbar is too small to
display the sliders. The slider shows instead in a popup which
fails to locate the device view model.

Fix that.

Fixes: 243768092
Test: N/A
Change-Id: I7f558884b446aea6aa5fb3f23cc4d3eabcf2068b
(cherry picked from commit bd661f267976a60fa245b4cf3dbf53c7313e3895)