Use State For showLogcat() Instead of selectDevice()
selectDevice() was never working because at the time that it's called, the device combo doesn't yet have the device. Rather than wait for the device combo to populate, we can take advantage of the LogcatPanelConfig state which already works.

Some architectural changes needed to be made in order to make this testable. Specifically, we need properly propagate the AdbSession object to all relevant components.

Removed the test and changes required to make the test work because these changes result in merge conflicts.

Fixes: 247728244
Test: Removed. It's tested on main but would result in merge conflicts here.
Change-Id: I5da5a3053567a4dbf4a5844945614b5b307ad80f