[platform] IDEA-305295 Fix visibility icons for project view panes
For some reason, only the Project View and Android Project View
explicitly implemented the ProjectViewSettings interface and its
isShowVisibilityIcons() method. This lead to a UI inconsistency:
when the user checks the corresponding option, it has no effect
on other panes.

Fix by pushing the code up to the ProjectTreeStructure class that
is a common ancestor for all project panes. Now the Show Visibility
Icons option works consistently for all panes.

GitOrigin-RevId: 9be9f60fedf122f91e5f50f2d4354fa64c806607